Add a license file and only only ship necessary libs in the gem
This adds a license file for services (like github) that parse out
licenses from a license file. It also stips the gemspec down to only
ship the necessary files in the gem artifact. There's no need for the
test files in the gem artifact and they make ruby installs larger.

Signed-off-by: Tim Smith <tsmith@chef.io>